replace download function from singleton to viewmodel
